CSA-55 added CoffeeShopMenuController tests for adding & editing items including with zero & negative prices and empty names, resolved merge conflicts with previous branch - updated code

+++ src/main/java/org/loose/fis/sre/controllers/CoffeeShopMenuClientController.java
@@ -72,7 +72,7 @@ public class CoffeeShopMenuClientController {
         newStage.show();
     }
 
-    public void createNewItemContainer(String name, String description, String drinkVolume, int price, CoffeeShopMenuItem item) {
+    public void createNewItemContainer(String name, String description, String drinkVolume, float price, CoffeeShopMenuItem item) {
         HBox newHBox = new HBox();
         AnchorPane newPanelContent = new AnchorPane();
