+++ src/main/java/org/loose/fis/sre/services/CoffeeShopMenuItemService.java
@@ -18,7 +18,7 @@ import static org.loose.fis.sre.services.FileSystemService.getPathToFile;
 
 public class CoffeeShopMenuItemService {
 
-    public static void addMenuItem(String name, String description, String drinkVolume, int price) throws InvalidIdException, MenuItemAlreadyExistsException {
+    public static void addMenuItem(String name, String description, String drinkVolume, float price) throws InvalidIdException, MenuItemAlreadyExistsException {
         checkMenuItemDoesNotAlreadyExist(name);
         CoffeeShopMenuItem[] newMenuItems = new CoffeeShopMenuItem[CoffeeShopMenuController.getCurrentCoffeeShop().getMenuItemsNumber() + 1];
         int count = 0;
@@ -55,7 +55,7 @@ public class CoffeeShopMenuItemService {
         CoffeeShopService.modifyCoffeeShop(CoffeeShopMenuController.getCurrentCoffeeShop());
     }
 
-    public static void modifyMenuItem(String oldName, String name, String description, String drinkVolume, int price) throws MenuItemAlreadyExistsException {
+    public static void modifyMenuItem(String oldName, String name, String description, String drinkVolume, float price) throws MenuItemAlreadyExistsException {
         if(!oldName.equals(name)) {
             removeMenuItem(oldName, description, drinkVolume);
             addMenuItem(name, description, drinkVolume, price);
